feat: Add standalone and test commands to forest CLI
